requirements.txt
django==5.0.7
djangorestframework==3.15.2
djoser==2.2.3 
gunicorn==22.0.0 
dj-database-url==2.2.0
whitenoise==6.7.0 
psycopg2-binary==2.9.9
mysqlclient==2.1.1
